+++ aura_ui/lib/aura_ui.dart
@@ -42,3 +42,4 @@ export 'src/texts/sparkle_text.dart';
 
 export 'src/buttons/buttons.dart';
 export 'src/toast/toasts.dart';
+export 'src/dialogs/dialogs.dart';